♻️ Refactor(sync): DB 직접 접근 제거, Repository 패턴으로 통일
- metadata-sync.ts: Drizzle 직접 호출 → CategoryRepository.rebuild(),
  FolderRepository(getAll/upsert/ensureFolder) 사용
- sync-strategies.ts: performFullSync 내 Drizzle 직접 호출
  → PostRepository(getAllForSync/update/create/deactivateMissing) 사용
- sync-github.ts: retitleExistingPosts → PostRepository.retitleAll() 위임
- post-sync.ts: getDb import 경로 @/db 로 변경
